therapist-agent/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
faiss-cpu
openai
pandas
python-dotenv
